requirements.txt
asgiref==3.5.0
content==0.0.1
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==12.0.1
Markdown==3.3.6
mixer==7.2.2
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
todo==0.1
total==0.0.7
